fixed some more functions & added more examples
